LPS-57740 move to module
