+++ modules/portal/portal-background-task-service/src/com/liferay/portal/background/task/model/impl/BackgroundTaskImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.model.impl;
+package com.liferay.portal.background.task.model.impl;
 
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskConstants;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskExecutor;
