+++ portal-impl/src/META-INF/model-listener-spring.xml
@@ -7,7 +7,6 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
 >
-	<bean id="com.liferay.portal.model.BackgroundTaskModelListener" class="com.liferay.portal.model.BackgroundTaskModelListener" />
 	<bean id="com.liferay.portal.model.LayoutModelListener" class="com.liferay.portal.model.LayoutModelListener" />
 	<bean id="com.liferay.portal.model.LayoutSetModelListener" class="com.liferay.portal.model.LayoutSetModelListener" />
 	<bean id="com.liferay.portal.model.PortletPreferencesModelListener" class="com.liferay.portal.model.PortletPreferencesModelListener" />
