+++ portal-impl/src/portal.properties
@@ -5611,7 +5611,6 @@
     # This property is not read by the portal except for portal properties
     # overridden by liferay-hook.xml. It remains here only as a reference.
     #
-    #value.object.listener.com.liferay.portal.model.BackgroundTask=com.liferay.portal.model.BackgroundTaskModelListener
     #value.object.listener.com.liferay.portal.model.Contact=com.liferay.portal.model.ContactModelListener
     #value.object.listener.com.liferay.portal.model.Layout=com.liferay.portal.model.LayoutModelListener,com.liferay.portal.service.impl.LayoutSetPrototypeLayoutModelListener
     #value.object.listener.com.liferay.portal.model.LayoutSet=com.liferay.portal.model.LayoutSetModelListener,com.liferay.portal.service.impl.LayoutSetPrototypeLayoutSetModelListener
